chore: split IconWrapper (#43775)
* chore: split icon-wrapper

* chore: add fallback

* chore: add fallback
